refactored panel in dashboard
