+++ src/sql/parts/dashboard/tabs/dashboardHomeTab.css
@@ -3,16 +3,8 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-dashboard-widget-tab {
+ dashboard-home-tab {
 	height: 100%;
 	width: 100%;
-}
-
-.scroll-container {
-	height: 100%;
-	width: 100%;
-}
-
-.scrollable {
-	position: relative;
+	display: block;
 }
